+++ src/main/java/dao/dao/UserDao.java
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 import java.util.List;
 
 public abstract class UserDao extends AbstractDao {
-    public abstract User getUserById(int id) throws SQLException;
+    public abstract User getUserById(Long id) throws SQLException;
 
     public abstract List<User> getAllUsers() throws SQLException;
